consolidate onboarding between mobile app and extension (#2807)
* init consolidate onboarding

* wip consolidate it all

* fix types

* bleh

* removes code

* consolidate between the two

* fix types

* wip

* updates

* clean up functions

* updates
