+++ packages/app-extension/src/components/Onboarding/pages/OnboardAccount.tsx
@@ -1,20 +1,16 @@
 import { useEffect, useState } from "react";
 import type {
-  Blockchain,
   KeyringType,
   SignedWalletDescriptor,
   WalletDescriptor,
 } from "@coral-xyz/common";
+import { getCreateMessage } from "@coral-xyz/common";
 import {
-  getBlockchainFromPath,
-  getCreateMessage,
-  UI_RPC_METHOD_FIND_WALLET_DESCRIPTOR,
-  UI_RPC_METHOD_SIGN_MESSAGE_FOR_PUBLIC_KEY,
-} from "@coral-xyz/common";
-import { useBackgroundClient } from "@coral-xyz/recoil";
-import { ethers } from "ethers";
+  useBackgroundClient,
+  useOnboarding,
+  useSignMessageForWallet,
+} from "@coral-xyz/recoil";
 
-import { useSignMessageForWallet } from "../../../hooks/useSignMessageForWallet";
 import { useSteps } from "../../../hooks/useSteps";
 import { CreatePassword } from "../../common/Account/CreatePassword";
 import { ImportWallets } from "../../common/Account/ImportWallets";
@@ -31,8 +27,6 @@ import { InviteCodeForm } from "./InviteCodeForm";
 import { KeyringTypeSelector } from "./KeyringTypeSelector";
 import { NotificationsPermission } from "./NotificationsPermission";
 import { UsernameForm } from "./UsernameForm";
-
-const { base58 } = ethers.utils;
 
 export const OnboardAccount = ({
   onWaiting,
@@ -51,105 +45,55 @@ export const OnboardAccount = ({
 }) => {
   const background = useBackgroundClient();
   const { step, nextStep, prevStep } = useSteps();
-  const [inviteCode, setInviteCode] = useState<string | undefined>(undefined);
-  const [username, setUsername] = useState<string | null>(null);
-  const [action, setAction] = useState<"create" | "import">();
-  const [keyringType, setKeyringType] = useState<KeyringType | null>(null);
-  const [blockchain, setBlockchain] = useState<Blockchain | null>(null);
-  const [password, setPassword] = useState<string | null>(null);
-  const [mnemonic, setMnemonic] = useState<string | undefined>(undefined);
   const [openDrawer, setOpenDrawer] = useState(false);
-  const [signedWalletDescriptors, setSignedWalletDescriptors] = useState<
-    Array<SignedWalletDescriptor>
-  >([]);
-
+  const { onboardingData, setOnboardingData, handleSelectBlockchain } =
+    useOnboarding();
+  const {
+    inviteCode,
+    username,
+    action,
+    keyringType,
+    password,
+    mnemonic,
+    blockchain,
+    signedWalletDescriptors,
+    selectedBlockchains,
+  } = onboardingData;
   const signMessageForWallet = useSignMessageForWallet(mnemonic);
-
-  const selectedBlockchains = [
-    ...new Set(
-      signedWalletDescriptors.map((s) =>
-        getBlockchainFromPath(s.derivationPath)
-      )
-    ),
-  ];
 
   useEffect(() => {
     // Reset blockchain keyrings on certain changes that invalidate the addresses
-    // and signatures that they might contain
-    // e.g. user has navigated backward through the onboarding flow
-    setSignedWalletDescriptors([]);
+    setOnboardingData({
+      signedWalletDescriptors: [],
+    });
   }, [action, keyringType, mnemonic]);
-
-  const handleBlockchainClick = async (blockchain: Blockchain) => {
-    if (selectedBlockchains.includes(blockchain)) {
-      // Blockchain is being deselected
-      setBlockchain(null);
-      setSignedWalletDescriptors(
-        signedWalletDescriptors.filter(
-          (s) => getBlockchainFromPath(s.derivationPath) !== blockchain
-        )
-      );
-    } else {
-      // Blockchain is being selected
-      if (keyringType === "ledger" || action === "import") {
-        // If wallet is a ledger, step through the ledger onboarding flow
-        // OR if action is an import then open the drawer with the import accounts
-        // component
-        setBlockchain(blockchain);
-        setOpenDrawer(true);
-      } else if (action === "create") {
-        const walletDescriptor = await background.request({
-          method: UI_RPC_METHOD_FIND_WALLET_DESCRIPTOR,
-          params: [blockchain, 0, mnemonic],
-        });
-        const signature = await background.request({
-          method: UI_RPC_METHOD_SIGN_MESSAGE_FOR_PUBLIC_KEY,
-          params: [
-            blockchain,
-            walletDescriptor.publicKey,
-            base58.encode(
-              Buffer.from(getCreateMessage(walletDescriptor.publicKey), "utf-8")
-            ),
-            [mnemonic, [walletDescriptor.derivationPath]],
-          ],
-        });
-        setSignedWalletDescriptors([
-          ...signedWalletDescriptors,
-          {
-            ...walletDescriptor,
-            signature,
-          },
-        ]);
-      }
-    }
-  };
 
   const steps = [
     <InviteCodeForm
       onClickWaiting={onWaiting}
       onClickRecover={onRecover}
       onSubmit={(inviteCode) => {
-        setInviteCode(inviteCode);
+        setOnboardingData({ inviteCode });
         nextStep();
       }}
     />,
     <UsernameForm
       inviteCode={inviteCode!}
       onNext={(username) => {
-        setUsername(username);
+        setOnboardingData({ username });
         nextStep();
       }}
     />,
     <CreateOrImportWallet
       onNext={(action) => {
-        setAction(action);
+        setOnboardingData({ action });
         nextStep();
       }}
     />,
     <KeyringTypeSelector
-      action={action!}
+      action={action}
       onNext={(keyringType: KeyringType) => {
-        setKeyringType(keyringType);
+        setOnboardingData({ keyringType });
         nextStep();
       }}
     />,
@@ -160,7 +104,7 @@ export const OnboardAccount = ({
             readOnly={action === "create"}
             buttonLabel={action === "create" ? "Next" : "Import"}
             onNext={(mnemonic) => {
-              setMnemonic(mnemonic);
+              setOnboardingData({ mnemonic });
               nextStep();
             }}
           />,
@@ -168,14 +112,22 @@ export const OnboardAccount = ({
       : []),
     <BlockchainSelector
       selectedBlockchains={selectedBlockchains}
-      onClick={handleBlockchainClick}
+      onClick={async (blockchain) => {
+        await handleSelectBlockchain({
+          blockchain,
+          background,
+          onSelectImport: () => {
+            setOpenDrawer(true);
+          },
+        });
+      }}
       onNext={nextStep}
     />,
     ...(!isAddingAccount
       ? [
           <CreatePassword
             onNext={(password) => {
-              setPassword(password);
+              setOnboardingData({ password });
               nextStep();
             }}
           />,
@@ -221,17 +173,20 @@ export const OnboardAccount = ({
         {keyringType === "ledger" ? (
           <HardwareOnboard
             blockchain={blockchain!}
-            action={action!}
+            // @ts-expect-error not assignable to type string ...
+            action={action}
             signMessage={(publicKey: string) => getCreateMessage(publicKey)}
             signText={`Sign the message to authenticate with Backpack.`}
+            onClose={() => setOpenDrawer(false)}
             onComplete={(signedWalletDescriptor: SignedWalletDescriptor) => {
-              setSignedWalletDescriptors([
-                ...signedWalletDescriptors,
-                signedWalletDescriptor,
-              ]);
+              setOnboardingData({
+                signedWalletDescriptors: [
+                  ...signedWalletDescriptors,
+                  signedWalletDescriptor,
+                ],
+              });
               setOpenDrawer(false);
             }}
-            onClose={() => setOpenDrawer(false)}
           />
         ) : (
           <ImportWallets
@@ -245,13 +200,15 @@ export const OnboardAccount = ({
                 walletDescriptor,
                 getCreateMessage(walletDescriptor.publicKey)
               );
-              setSignedWalletDescriptors([
-                ...signedWalletDescriptors,
-                {
-                  ...walletDescriptor,
-                  signature,
-                },
-              ]);
+              setOnboardingData({
+                signedWalletDescriptors: [
+                  ...signedWalletDescriptors,
+                  {
+                    ...walletDescriptor,
+                    signature,
+                  },
+                ],
+              });
               setOpenDrawer(false);
             }}
           />
