+++ packages/app-mobile/src/lib/index.ts
@@ -1,0 +1,10 @@
+export function maybeRender(
+  condition: boolean,
+  fn: () => JSX.Element
+): JSX.Element | null {
+  if (condition) {
+    return fn() as JSX.Element;
+  }
+
+  return null;
+}
