+++ packages/recoil/src/context/index.tsx
@@ -1,3 +1,4 @@
 export * from "./NavEphemeral";
 export * from "./Notifications";
+export * from "./OnboardingProvider";
 export * from "./Swap";
