+++ packages/recoil/src/hooks/index.tsx
@@ -14,6 +14,7 @@ export * from "./useLoader";
 export * from "./usePriceData";
 export * from "./usePrimaryWallets";
 export * from "./useRecentNotifications";
+export * from "./useSignMessageForWallet";
 export * from "./useTokenMetadata";
 export * from "./useTotalBalance";
 export * from "./useTransactionData";
